Updates to ScanServerAttempts
As a follow-up to #2880:

* Remove unused and undocumented endTime in ScanAttempt API
* Separate out more inner classes/interfaces into their own files
* Simplify concurrency by synchronizing on the attempts and preparing
  the snapshot using immutable collections
* IDE changes also converted some anonymous inner classes to lambdas
